Add dataverk_init script and test and add to setup.py
dataverk_init is a CLI scripts that creates new datasett projects.
It asks the user for input and handles the creation of folders and files
like LICENCE and CODEOWNERS for a new prosject. Finaly it initializes a
git repo.

[TODO]
create a new github repo under navikt
more tests
better error handling